This version reads a collide .dae file and is now able to parse individual nodes in the file. the nodes are created to each have their own local geometry with a global to local transformation matrix. If this isn't handled correctly then 'group' or components within a dae file do not get translated to the correct world coordinates.
The program can write out the parsed file as a .ply file that can be read by mesh viewers such as meshlab.

+++ readCollada/colladainterface.h
@@ -49,7 +49,7 @@ struct ColGeom {
     int index_count ;
     unsigned short* indices ;
     float scaling ;
-    float transform[16];
+    double transform[16];
     std::string material ;
 };
